pyproject.toml
[project]
name = "autotrade"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "backtesting>=0.6.5",
    "ccxt>=4.5.0",
    "pandas>=2.2.3",
    "ta>=0.11.0",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.0.1",
]
